backend: мелкие фиксы

+++ backend/handlers/day_handlers.go
@@ -21,7 +21,7 @@ func CreateDay(c *fiber.Ctx) error {
 
     if err := c.BodyParser(day); err != nil { return err }
 
-	id, err := database.CreateDay(int16(day.DayNumber), string(day.DayType))
+	id, err := database.CreateDay(day.DayNumber, string(day.DayType))
 	if err != nil { return err }
 	day.DayID = id
